#128 move benchmark from test to app to avoid the catch2 req
